requirements.txt
alembic==1.16.1
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
click==8.2.1
fastf1>=3.5.3
Flask==3.1.1
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
libpass==1.9.1.post0
matplotlib>=3.7.0
matplotlib-inline==0.1.7
numpy==2.2.6
pandas>=2.0.0
passlib==1.7.4
pytest-flask==1.3.0
python-dotenv==1.1.0
requests==2.32.3
requests-cache==1.2.1
types-requests==2.32.0.20250515
 